Add mastercf_tail option
This option lets you append anything to the end of your master.cf. This
is particularly useful if you have some custom transports, or other
configurations that you need to make that aren't part of the module, and
don't make sense to add to the module because they are too site-specific.