Allowed mynetworks to be an array
Updatd main.cf to accept mynetworks as either a string or an array of
strings.
Updated mta.cf to accept a string or an array of strings. If an array is
passed it will be join to a string before validating.

The purpose of this it be make larger lists of trusted networks easier
to read and make it easier to locate changes to the list in a diff.

In the following example it would be troublesome/time consuming to find a single
changed network if it was a long string instead of an array.
Example:

  class { 'postfix':
    relayhost     => 'mail.example.com',
    smtp_listen   => '0.0.0.0',
    mydestination => '$myorigin, myapp.example.com',
    mta           => true,
    mynetworks    => [
                       '192.0.2.0/31',
                       '192.0.2.64/31',
                       '192.0.2.80/31',
                       '192.0.2.100/31',
                       '192.0.2.120/31',
                       '192.0.2.160/31',
                       '192.0.2.188/31',
                       '192.0.2.200/31',
                       '192.0.2.240/31',
                       '198.51.100.0/31',
                       '198.51.100.64/31',
                       '198.51.100.80/31',
                       '198.51.100.100/31',
                       '198.51.100.120/31',
                       '198.51.100.160/31',
                       '198.51.100.188/31',
                       '198.51.100.200/31',
                       '198.51.100.240/31',
                       '203.0.113.0/31',
                       '203.0.113.64/31',
                       '203.0.113.80/31',
                       '203.0.113.100/31',
                       '203.0.113.120/31',
                       '203.0.113.160/31',
                       '203.0.113.188/31',
                       '203.0.113.200/31',
                       '203.0.113.240/31',
                     ],
  }