Schleuder: support explicitly setting use_schleuder to 3.
A Boolean "true" value will thus mean "assume the latest upstream Schleuder
release". For now, that's 3.x so setting this parameter to 3 is equivalent
to setting it to true. But once Schleuder 4 is out, this may change,
so better provide users a way to hard-code the fact they have Schleuder 3.

Thanks to @raphink for the suggestion!