Schleuder: port invocation syntax to Schleuder 3. (#234)
* Schleuder: port invocation syntax to Schleuder 3.

Schleuder 2.x has been deprecated for a while. It's not maintained
anymore, has obsolete dependencies, and has been superseded by
Schleuder 3.

The new syntax is documented on
https://schleuder.org/schleuder/docs/server-admins.html#postfix

* compatibility with Schleuder 2 can be used this way:

  use_schleuder => 2

* support explicitly setting use_schleuder to 3.

A Boolean "true" value will thus mean "assume the latest upstream Schleuder
release". For now, that's 3.x so setting this parameter to 3 is equivalent
to setting it to true. But once Schleuder 4 is out, this may change,
so better provide users a way to hard-code the fact they have Schleuder 3.