Fix hardcoded map path
The relationship definition for the map path in `Postfix::Transport` was
hardcoded, so that any non-standard `$file` parameter did not create an
automatic refresh of the map, even if it is defined as a
`Postfix::Hash`.

Fixed be replacing the hardcoded string with the `$file` variable, which
means no change with expected behaviour.