Add FreeBSD support (#288)
* Do not hardcode /etc/postfix

In order to bring support for FreeBSD, do not use the hardcoded
/etc/postfix path for the directory containing Postfix configuration:
FreeBSD ports are installed with a /usr/local/ prefix (by default) so
the configuration files on FreeBSD are stored in the
/usr/local/etc/postfix directory.

* Add support for FreeBSD

* Make it possible to tune the "root" group

FreeBSD does not have a "root" group.  The corresponding group is named
"wheel".

Allow to setup a custom "root_group", and adjust FreeBSD configuration
to set it to "wheel".

* Adjust the test suite so that it pass on FreeBSD

* Do not depend on $postfix::* before including postfix

Some resources parameters depend on the value of variable from the
postfix class.  Ensure these values are substitued only after including
postfix.

* Move default values from hiera to init.pp

These values are system-dependent, but this helps seeing the usual
default value when genering references.

* Move $manage_mailname parameter

* Explicitely mark internal classes as private

* Remove redundant postfix::params inclusion

Co-authored-by: Raphaël Pinson <github+aem1eeshi1@raphink.net>

Co-authored-by: Raphaël Pinson <github+aem1eeshi1@raphink.net>